# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bsllmner-viewer API",
    "description": "Queries, aggregations, and exports over ontology-mapped BioSample annotations. The conventions follow the DDBJ Search API.\n\nStart with [/llms.txt](/llms.txt): condition examples and recipes for common tasks. [/llms-full.txt](/llms-full.txt) has the rules that span operations: the condition language, the descendants of terms, counting units and populations, self-exclusion, default elements, expected counts, export formats, errors, limits, and caching with `ETag` and `If-None-Match`. A description in this document names the heading of the rule that it relies on.",
    "contact": {
      "name": "BioData Science Initiative",
      "url": "https://github.com/dbcls/bsllmner-viewer"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/entries/{type}": {
      "get": {
        "tags": [
          "Entries"
        ],
        "summary": "List the BioSamples that match a condition",
        "description": "Lists the BioSamples that match `q`, one page at a time. `pagination.total` is the count of `q` in the BioSample unit. Each item lists the SRA Experiments of the BioSample that match `q`. An entry type that is not `biosample` gets 404. The Samples view of the UI uses this operation. See \"Entries\" in /llms-full.txt.",
        "operationId": "listEntries",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntryType",
              "description": "Entry type. Only `biosample` has entries"
            },
            "description": "Entry type. Only `biosample` has entries"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt",
              "examples": [
                "disease:\"MONDO:0007254\"",
                "cell_line:\"CVCL:0030\" AND library_strategy:ATAC-seq",
                "organism_id:9606 AND date_published:[2015-01-01 TO 2020-12-31]",
                "disease_status:unmapped",
                "\"breast cancer\" AND NOT organism_id:10090"
              ],
              "title": "Q"
            },
            "description": "The condition, in the condition language: `field:value`, `field:\"phrase\"`, `field:[a TO b]`, keywords without a field, `AND`, `OR`, `NOT`, and parentheses. If `q` is omitted or empty, the condition is the whole population. The value of an annotation term field is a term ID. `GET /api/terms` finds term IDs. A term ID matches the term and its descendants. The value of `<field>_status` is a status group: `mapped`, `unmapped`, or `no_value`. The value of `library_strategy` is a target assay of `GET /api/dataset`. A keyword matches the searchable text of a BioSample. A keyword in the form of an accession matches the entries that have the accession instead. See \"Condition DSL\" in /llms-full.txt"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number, starting at 1. A page after the last page has empty `items`",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, starting at 1. A page after the last page has empty `items`"
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page, from 1 to 100",
              "default": 25,
              "title": "Perpage"
            },
            "description": "Items per page, from 1 to 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (the request does not match the operation)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. The slug that ends `type` is one of:\n\n- `unexpected-token`: The condition does not parse, is longer than 4096 characters, or is blank where a condition is required, or a keyword has a wildcard. Parse, select, and keyword also return this slug if the condition that they return would be longer than 4096 characters. Fix the condition at the column that `detail` gives. Quote a value that has a colon or a space. See \"Grammar\" in /llms-full.txt.\n- `unknown-field`: A condition, a clause, or a parameter (`field`, `row`, `col`) names a field that the dataset does not have. Use a field that `dslFields` of `GET /api/dataset` lists.\n- `invalid-date-format`: A value of `date_published` in the form `YYYY-MM-DD` is not a calendar date, such as `2020-13-45`, or a bound of a range is not a date in that form. Write the date as `YYYY-MM-DD`, such as `2020-12-31`. See \"Fields\" in /llms-full.txt.\n- `invalid-operator-for-field`: The form of a value does not fit the field, such as a range or a date on a field other than `date_published`, a value with a wildcard, or a value of `date_published` that is neither a date in the form `YYYY-MM-DD` nor a range (`2020-1-1`, a quoted date, or `\"\"`). Use `field:value` for the field, `date_published:YYYY-MM-DD` for a date, or `date_published:[a TO b]` for a range. See \"Fields\" in /llms-full.txt.\n- `invalid-value`: A value is not allowed for the field: a status that is not a status group, an assay that the dataset does not have, a term value that is not a term ID, an organism that is not a number, a keyword without a letter or a digit, or keywords over their limits. Use the values that `detail` lists. Find a term ID with `GET /api/terms`. See \"Fields\" and \"Keywords\" in /llms-full.txt.\n- `nest-depth-exceeded`: Groups nest deeper than 5 levels, or the condition has more than 512 nodes. Flatten the groups or split the condition. See \"Limits\" in /llms-full.txt.\n- `missing-value`: A clause on a field other than `date_published` has an empty value. Give the clause a value. See \"Fields\" in /llms-full.txt.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable. The slug that ends `type` is one of:\n\n- `server-busy`: The api has no free slot for the request after waiting, or too many requests wait already. Wait for the `Retry-After` seconds and send the request again.\n- `query-timeout`: The query took longer than the time limit of a request and was stopped. Narrow the condition. The same request does not finish by itself.\n- `query-too-large`: The query needs more memory or temporary disk space than a worker may use and was stopped. Narrow the condition, or ask for fewer elements.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/entries/biosample/{accession}": {
      "get": {
        "tags": [
          "Entries"
        ],
        "summary": "Get a BioSample with its annotations and evidence",
        "description": "Returns one BioSample with its original metadata, annotations with evidence, SRA Experiments, and BioProjects. The BioSample does not have to be in the population: `inPopulation` of each SRA Experiment tells whether the SRA Experiment is in the population. `run` is the name of the bsllmner-mk2 run that analyzed the BioSample, not an SRA Run. An accession that is not that of a BioSample gets 404. Find the BioSample of another accession with the accession as a keyword in `q`. See \"Entries\" in /llms-full.txt.",
        "operationId": "getEntry",
        "parameters": [
          {
            "name": "accession",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "description": "BioSample accession, such as `SAMN14864678`",
              "title": "Accession"
            },
            "description": "BioSample accession, such as `SAMN14864678`"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (the request does not match the operation)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Clause-Output": {
        "$ref": "#/components/schemas/ClauseJson"
      },
...
      "DatasetVersionRef": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name of the dataset version"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat",
            "description": "Time when the store was built, ISO 8601"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "The model that the bsllmner-mk2 runs of the dataset used"
          },
          "digest": {
            "type": "string",
            "title": "Digest",
            "description": "Short hash of the full version information"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "createdAt",
          "model",
          "digest"
        ],
        "title": "DatasetVersionRef",
        "description": "Identifies the store a response was computed from. Full details are returned by the dataset endpoint."
      },
...
      "EntriesResponse": {
        "properties": {
          "datasetVersion": {
            "$ref": "#/components/schemas/DatasetVersionRef"
          },
          "q": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          },
          "type": {
            "$ref": "#/components/schemas/EntryType"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntryItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "datasetVersion",
          "q",
          "type",
          "pagination",
          "items"
        ],
        "title": "EntriesResponse"
      },
      "EntryAnnotation": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Annotation field name"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "The extracted value; null when none was extracted"
          },
          "status": {
            "$ref": "#/components/schemas/Status",
            "description": "Annotation status. The UI shows `mapped_exact` as \"Exact match\", `mapped_selected` as \"LLM selected\" (LLM: large language model), `unmapped_no_candidate` as \"No candidate\", `unmapped_rejected` as \"Rejected\", `not_stated` as \"Not stated\", and `extraction_failed` as \"Extraction failed\". See \"Annotation status\" in /llms-full.txt for the meanings"
          },
          "termId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Termid",
            "description": "ID of the mapped term; null when the value has no term"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "Label of the mapped term; null when the value has no term"
          },
          "clauses": {
            "items": {
              "$ref": "#/components/schemas/Clause-Output"
            },
            "type": "array",
            "title": "Clauses",
            "description": "The clause on the field and the term; empty without a term"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "value",
          "status",
          "termId",
          "label",
          "clauses",
          "evidence"
        ],
        "title": "EntryAnnotation"
      },
...
      "EntryItem": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier",
            "description": "BioSample accession"
          },
          "type": {
            "$ref": "#/components/schemas/EntryType"
          },
          "experiments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Experiments",
            "description": "SRA Experiments of the BioSample that match the condition"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "organism": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organism"
              },
              {
                "type": "null"
              }
            ]
          },
          "libraryStrategy": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Librarystrategy",
            "description": "Assays of the matching SRA Experiments"
          },
          "bioprojects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bioprojects",
            "description": "Accessions of the BioProjects of the BioSample"
          },
          "datePublished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datepublished",
            "description": "Publication date, `YYYY-MM-DD`"
          },
          "chipAtlas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chipatlas",
            "description": "Genome assemblies under which ChIP-Atlas processed the SRA Experiments"
          },
          "annotations": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/AnnotationValue"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Annotations",
            "description": "Annotations per annotation field name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identifier",
          "type",
          "experiments",
          "title",
          "organism",
          "libraryStrategy",
          "bioprojects",
          "datePublished",
          "chipAtlas",
          "annotations"
        ],
        "title": "EntryItem"
      },
      "EntryResponse": {
        "properties": {
          "datasetVersion": {
            "$ref": "#/components/schemas/DatasetVersionRef"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier",
            "description": "BioSample accession"
          },
          "type": {
            "type": "string",
            "const": "biosample",
            "title": "Type"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "organism": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organism"
              },
              {
                "type": "null"
              }
            ]
          },
          "datePublished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datepublished"
          },
          "run": {
            "type": "string",
            "title": "Run",
            "description": "Name of the bsllmner-mk2 run that analyzed the BioSample. It is not an SRA Run"
          },
          "metadata": {
            "items": {
              "$ref": "#/components/schemas/MetadataItem"
            },
            "type": "array",
            "title": "Metadata",
            "description": "Original metadata in the order of the description, the record, and the attributes. `metadataIndex` of an evidence is a position in this list"
          },
          "annotations": {
            "items": {
              "$ref": "#/components/schemas/EntryAnnotation"
            },
            "type": "array",
            "title": "Annotations"
          },
          "experiments": {
            "items": {
              "$ref": "#/components/schemas/EntryExperiment"
            },
            "type": "array",
            "title": "Experiments"
          },
          "bioprojects": {
            "items": {
              "$ref": "#/components/schemas/EntryBioProject"
            },
            "type": "array",
            "title": "Bioprojects"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "datasetVersion",
          "identifier",
          "type",
          "title",
          "organism",
          "datePublished",
          "run",
          "metadata",
          "annotations",
          "experiments",
          "bioprojects"
        ],
        "title": "EntryResponse"
      },
      "EntryType": {
        "type": "string",
        "const": "biosample"
      },
...
      "MetadataItem": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/MetadataKind"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name to show: a description name, an attribute name, or a short name of a path"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "harmonizedName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harmonizedname",
            "description": "Harmonized name of an attribute; null for the other kinds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "name",
          "value",
          "harmonizedName"
        ],
        "title": "MetadataItem"
      },
      "MetadataKind": {
        "type": "string",
        "enum": [
          "description",
          "attribute",
          "record"
        ]
      },
...
      "Organism": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier",
            "description": "NCBI Taxonomy ID"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Name of the organism; null when no name is known"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identifier",
          "name"
        ],
        "title": "Organism",
        "description": "An organism as the NCBI Taxonomy ID and the name."
      },
      "Pagination": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "perPage": {
            "type": "integer",
            "title": "Perpage"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Count of all items of the list in its counting unit, not only of this page"
          },
          "hasNext": {
            "type": "boolean",
            "title": "Hasnext",
            "description": "Whether a page after this one has items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "perPage",
          "total",
          "hasNext"
        ],
        "title": "Pagination"
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "`about:blank` for an error that the HTTP status describes. Otherwise a URI that ends with the slug of the rule that the request broke. The description of the 400 response of each operation lists its slugs."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "The HTTP status phrase"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance",
            "description": "The path of the request"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 time in UTC"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid",
            "description": "The value of the `X-Request-ID` header of the response"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "timestamp",
          "requestId"
        ],
        "title": "ProblemDetails",
        "description": "RFC 7807 problem details of an error response."
      },
...
      "Status": {
        "type": "string",
        "enum": [
          "mapped_exact",
          "mapped_selected",
          "unmapped_no_candidate",
          "unmapped_rejected",
          "not_stated",
          "extraction_failed"
        ]
      },
...
    }
  },
...
}